refactor: start to fix typing issues with store
In order to have a passing test and build step, the code has to work. Perhaps the older typescript was more lax? Perhaps I got lucky. It used to build, but now it doesn't. Claude's helping me tighten it up, but we're not done yet. The core issues are around making reducers, stores and actions more generic.

I'm tempted to push on and make this more like modern redux-toolkit or redux-actions: where the state slices and action creators are all combined. I'd even like to have it such that action creators call dispatch directly. I think that'll make the code less verbose overall. But all this astronaut architecture is getting in the way of CI/CDing the project, so I'll stop.